Conflicts:
	cmake_targets/CMakeLists.txt
	cmake_targets/build_oai
	openair1/PHY/TOOLS/time_meas.h

+++ openair3/GTPV1-U/nw-gtpv1u/src/NwGtpv1u.c
@@ -59,6 +59,7 @@
 extern "C" {
 #endif
 
+//#define LOG_GTPU 1
 /*--------------------------------------------------------------------------*
  *                    P R I V A T E    F U N C T I O N S                    *
  *--------------------------------------------------------------------------*/
@@ -900,7 +901,7 @@ nwGtpv1uProcessUdpReq( NW_IN NwGtpv1uStackHandleT hGtpuStackHandle,
   case NW_GTP_END_MARKER:
 #if defined(LOG_GTPU) && LOG_GTPU > 0
     for(int i =1; i<= udpDataLen; i++){
-      printf("%02x ", udpData[i-1]);
+      printf("%02x ", (uint8_t)udpData[i-1]);
       if(i % 20 == 0)printf("\n");
     }
 #endif
